readonly'fy and remove coerce by fixing it in the builder:
The only usage for the builder (and thus the coerce) is the Runner
and the Config role. The Config role already returns the absolute
path, so the only need for the coerce is the Runner.

By making sure the builder in the Runner returns the absolute, we
no longer need to coerce the value.

There is no need for the "location" attribute to be readwrite
anymore.