move postponed_hooks to server from runner:
The runner was holding on to global postponed_hooks which would then
be inserted into the app (or new postponed_hooks from the app would
be added to it), but there was no point holding it in the runner. It
should have in fact been in the server. (one could claim that not even
there)

This commit moves the postponed_hooks to the server instead of the
runner. Also, in order to make this available, it exposes the runner
object using $Dancer2::runner or Dancer2->runner. Since it's already
global, might as well make it public so it would be easier to reach
the server.

This also removes the need to save a reference of the runner in the
server. This creates a circular reference that is hopefully not a
problem with the use of weak_ref which I hope we could eliminate in
the first place.